Finally opening a file in the buffer list!
I had to change some things to make this happen. First, I moved the mmap
line to main.c file. This kind of make sense: it's better to let the
main program to manage the memory used with the open files than to keep
all buffers always available in the memory, whithin the core library.

+++ src/core/include/getan_buffer.h
@@ -44,7 +44,7 @@ struct getan_buffer_cb {
 	getan_error (*init)(void **priv);
 	getan_error (*destroy)(void *priv);
 	getan_error (*call)(void *priv, unsigned int met, void *parm, size_t plen);
-	getan_error (*get)(void *priv, unsigned int attr, void *data, size_t *dlen);
+	getan_error (*get)(void *priv, unsigned int attr, void *data, int *dlen);
 	getan_error (*set)(void *priv, unsigned int attr, void *data, size_t dlen);
 };
